Move extensions to threepseat.ext
