+++ threepseat/ext/birthdays/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+from threepseat.ext.birthdays.commands import BirthdayCommands
