+++ threepseat/ext/rules/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+from threepseat.ext.rules.commands import RulesCommands
