+++ threepseat/ext/rules/exceptions.py
