+++ threepseat/ext/sounds/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+from threepseat.ext.sounds.commands import SoundCommands
