+++ threepseat/main.py
@@ -10,14 +10,14 @@ from typing import Any
 
 import threepseat
 from threepseat import config
-from threepseat.birthdays.commands import BirthdayCommands
 from threepseat.bot import Bot
-from threepseat.custom.commands import CustomCommands
+from threepseat.ext.birthdays import BirthdayCommands
+from threepseat.ext.custom import CustomCommands
+from threepseat.ext.reminders import ReminderCommands
+from threepseat.ext.rules import RulesCommands
+from threepseat.ext.sounds import SoundCommands
+from threepseat.ext.sounds.web import create_app
 from threepseat.logging import configure_logging
-from threepseat.reminders.commands import ReminderCommands
-from threepseat.rules.commands import RulesCommands
-from threepseat.sounds.commands import SoundCommands
-from threepseat.sounds.web import create_app
 
 
 logger = logging.getLogger(__name__)
